refactor(api): eliminate controller barrel re-exports and repoint contentHandler
