--- backend/api/src/controllers/admin/catalog/catalogBrandModelController.ts
@@ -1,2 +1,0 @@
-export * from './adminBrandController';
-export * from './adminModelController';
