--- backend/api/src/controllers/catalog/index.ts
@@ -1,6 +1,0 @@
-/**
- * Catalog SSOT bridge:
- * backend catalog routes proxy directly to core canonical controllers.
- */
-export * from '../admin/catalog';
-
