--- backend/api/src/utils/contentHandler.ts
@@ -1,1 +1,0 @@
-export * from './content-handler';
